rename files and updated packages
