+++ .vscode/settings.json
@@ -18,7 +18,8 @@
     "Modeline",
     "Prio",
     "schema",
-    "schemaservice"
+    "schemaservice",
+    "unzipper"
   ],
   "mochaExplorer.files": "server/test/**/*.ts",
   "mochaExplorer.require": "ts-node/register"
