+++ package.json
@@ -93,12 +93,12 @@
     "test-compile": "tsc -p ./"
   },
   "devDependencies": {
-    "@types/mocha": "8.2.2",
-    "@types/node": "^12.20.33",
+    "@types/mocha": "^10.0.10",
+    "@types/node": "^22.16.3",
     "@typescript-eslint/parser": "^4.16.1",
     "eslint": "^7.2.0",
     "merge-options": "^3.0.4",
-    "mocha": "^9.2.2",
+    "mocha": "^11.7.1",
     "node-ts": "^5.1.2",
     "rimraf": "^3.0.2",
     "ts-loader": "^8.3.0",
