+++ server/package.json
@@ -38,7 +38,7 @@
     "@types/ws": "^7.4.7",
     "@types/yaml": "1.9.7",
     "chai": "^4.3.6",
-    "mocha": "^9.2.1",
+    "mocha": "^11.7.1",
     "ts-node": "^10.9.1",
     "ts-node-dev": "^2.0.0",
     "typescript": "^4.7.4"
