+++ server/src/settings.ts
