+++ server/test/completion-sensors.test.ts
@@ -1,9 +1,8 @@
 import "mocha";
-import { expect } from "chai";
 import { getCompletionsFor, testCompletionHaveLabels } from "./shared";
 
 describe("completion-sensors", () => {
-  it("list sgp4x memebers", async () => {
+  it("list sgp4x members", async () => {
     const result = await getCompletionsFor(
       `
 sensor:
